# applications/AspNetWebApplication/openapi.json
{
"openapi": "3.0.1",
    "info": {
        "title": "Service ImageProcessor",
        "description": "Description for ImageProcessor service",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Uploads main page for application",
                "description": "Returns main page",
                "operationId": "GetMainPage",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/images/save": {
            "post": {
                "summary": "Saves sended image",
                "description": "Saves sended image, returns image_id",
                "operationId": "SaveImage",
                "parameters": [
                    {
                        "name": "table_element_type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ROW",
                                "COLUMN"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "image/*": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns image_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "image_id": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "image_id"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/images/{image_id}": {
            "get": {
                "summary": "Get image by image_id",
                "description": "Returns image by image_id",
                "operationId": "GetImage",
                "responses": {
                    "200": {
                        "description": "Returns image",
                        "content": {
                            "image/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/images/start_calculation": {
            "post": {
                "summary": "Start calculation of distance and similarity",
                "description": "Asynchronously starts calculation of distance and similarity",
                "operationId": "StartCalculation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImageIdsPair"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calculation started"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/images/compare": {
            "post": {
                "summary": "Compares images, calculates distance and similarity",
                "description": "Compares images, calculates distance and similarity, returns calculation progress indicator",
                "operationId": "CompareImages",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImageIdsPair"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Images comparsion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DistanceAndSimilarity"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/images/cancel_calculation": {
            "post": {
                "summary": "Cancels calculation of distance and simialrity",
                "description": "Cancels calculation of distance and simialrity",
                "operationId": "CancelCalculation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImageIdsPair"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calculation cancelled"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/images": {
            "get": {
                "summary": "Returns all image_ids from storage",
                "description": "Returns all image_ids from storage",
                "operationId": "GetImageIds",
                "responses": {
                    "200": {
                        "description": "All image_ids are returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageIdsByRowAndColumn"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "delete": {
                "summary": "Delete all images from table",
                "description": "Delete all images from table",
                "operationId": "DeleteImages",
                "responses": {
                    "200": {
                        "description": "All images are deleted"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "message"
                ]
            },
            "ImageIdsPair": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "first_image_id": {
                        "type": "string"
                    },
                    "second_image_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "first_image_id",
                    "second_image_id"
                ]
            },
            "DistanceAndSimilarity": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "distance": {
                        "type": "number"
                    },
                    "similarity": {
                        "type": "number"
                    },
                    "progress": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 1
                    }
                },
                "required": [
                    "progress"
                ]
            },
            "ImageIdsByRowAndColumn": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "row": {
                        "type": "array",
                        "description": "Image ids from table row",
                        "items": {
                            "description": "Image id",
                            "type": "string"
                        }
                    },
                    "column": {
                        "type": "array",
                        "description": "Image ids from table column",
                        "items": {
                            "description": "Image id",
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "row",
                    "column"
                ]
            }
        },
        "responses": {
            "400": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "404": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "500": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        }
    }
}